requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
mongoengine==0.13.0
pymongo==3.4.0
python-dateutil==2.6.1
requests==2.18.1
requests-cache==0.4.13
six==1.10.0
urllib3==1.21.1
